+++ datasophon-kubernetes/src/main/java/com/datasophon/kubernetes/strategy/KubernetesHbaseHandlerStrategy.java
@@ -71,7 +71,7 @@ public class KubernetesHbaseHandlerStrategy extends KubernetesAbstractHandlerStr
     }
 
     @Override
-    public void getConfig(Integer clusterId, String namespace, List<ServiceConfig> list) {
+    public void getConfig(Long clusterId, String namespace, List<ServiceConfig> list) {
         if (list == null || list.isEmpty()) {
             return;
         }
